+++ src/test/java/com/nhnacademy/couponapi/presentation/controller/CouponPolicyBookControllerTest.java
@@ -1,7 +1,10 @@
 package com.nhnacademy.couponapi.presentation.controller;
 
+import com.nhnacademy.couponapi.application.service.CouponPolicyBookService;
 import com.nhnacademy.couponapi.application.service.CouponPolicyService;
+import com.nhnacademy.couponapi.presentation.dto.request.CouponPolicyBookRequestDTO;
 import com.nhnacademy.couponapi.presentation.dto.request.CouponPolicyRequestDTO;
+import com.nhnacademy.couponapi.presentation.dto.response.CouponPolicyBookResponseDTO;
 import com.nhnacademy.couponapi.presentation.dto.response.CouponPolicyResponseDTO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -19,67 +22,84 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-class CouponPolicyControllerTest {
+class CouponPolicyBookControllerTest {
 
     @Mock
-    private CouponPolicyService couponPolicyService;
+    private CouponPolicyBookService couponPolicyBookService;
 
     @InjectMocks
-    private CouponPolicyController couponPolicyController;
+    private CouponPolicyBookController couponPolicyBookController;
 
-    private CouponPolicyRequestDTO couponPolicyRequestDTO;
-    private CouponPolicyResponseDTO couponPolicyResponseDTO;
+    private CouponPolicyBookRequestDTO couponPolicyBookRequestDTO;
+    private CouponPolicyBookResponseDTO couponPolicyBookResponseDTO;
 
     @BeforeEach
     void setUp() {
-        couponPolicyRequestDTO = new CouponPolicyRequestDTO(
-                "Test Policy",
-                new BigDecimal("10.00"),
-                null,
-                new BigDecimal("50.00"),
-                new BigDecimal("100.00"),
-                true
-        );
-
-        couponPolicyResponseDTO = CouponPolicyResponseDTO.builder()
-                .couponPolicyId(1L)
-                .couponPolicyName("Test Policy")
-                .couponPolicyDiscountValue(new BigDecimal("10.00"))
-                .couponPolicyCreatedAt(new Date())
-                .couponPolicyUpdatedAt(new Date())
-                .couponPolicyRate(null)
-                .couponPolicyMinOrderAmount(new BigDecimal("50.00"))
-                .couponPolicyMaxAmount(new BigDecimal("100.00"))
-                .couponPolicyDiscountType(true)
-                .build();
+        couponPolicyBookRequestDTO = new CouponPolicyBookRequestDTO(1L, 1L);
+        couponPolicyBookResponseDTO = new CouponPolicyBookResponseDTO(1L, 1L, 1L);
     }
 
     @Test
     public void testFindAll() {
-        when(couponPolicyService.findAllCouponPolicies()).thenReturn(List.of(couponPolicyResponseDTO));
+        when(couponPolicyBookService.findAllCouponPolicyBooks()).thenReturn(List.of(couponPolicyBookResponseDTO));
 
-        ResponseEntity<List<CouponPolicyResponseDTO>> response = couponPolicyController.findAll();
+        ResponseEntity<List<CouponPolicyBookResponseDTO>> response = couponPolicyBookController.findAll();
 
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
         assertThat(response.getBody()).hasSize(1);
-        assertThat(response.getBody().get(0).couponPolicyId()).isEqualTo(1L);
+        assertThat(response.getBody().get(0).couponPolicyBookId()).isEqualTo(1L);
 
-        verify(couponPolicyService, times(1)).findAllCouponPolicies();
+        verify(couponPolicyBookService, times(1)).findAllCouponPolicyBooks();
+    }
+
+    @Test
+    public void testFind() {
+        when(couponPolicyBookService.findCouponPolicyBookById(anyLong())).thenReturn(couponPolicyBookResponseDTO);
+
+        ResponseEntity<CouponPolicyBookResponseDTO> response = couponPolicyBookController.find(1L);
+
+        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
+        assertThat(response.getBody().couponPolicyBookId()).isEqualTo(1L);
+
+        verify(couponPolicyBookService, times(1)).findCouponPolicyBookById(1L);
     }
 
     @Test
     public void testCreate() {
-        when(couponPolicyService.createCouponPolicy(any(CouponPolicyRequestDTO.class))).thenReturn(couponPolicyResponseDTO);
+        when(couponPolicyBookService.createCouponPolicyBook(any(CouponPolicyBookRequestDTO.class))).thenReturn(couponPolicyBookResponseDTO);
 
-        ResponseEntity<CouponPolicyResponseDTO> response = couponPolicyController.create(couponPolicyRequestDTO);
+        ResponseEntity<CouponPolicyBookResponseDTO> response = couponPolicyBookController.create(couponPolicyBookRequestDTO);
 
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
-        assertThat(response.getBody().couponPolicyId()).isEqualTo(1L);
+        assertThat(response.getBody().couponPolicyBookId()).isEqualTo(1L);
 
-        verify(couponPolicyService, times(1)).createCouponPolicy(couponPolicyRequestDTO);
+        verify(couponPolicyBookService, times(1)).createCouponPolicyBook(couponPolicyBookRequestDTO);
     }
 
+    @Test
+    public void testUpdate() {
+        when(couponPolicyBookService.updateCouponPolicyBook(anyLong(), any(CouponPolicyBookRequestDTO.class))).thenReturn(couponPolicyBookResponseDTO);
+
+        ResponseEntity<CouponPolicyBookResponseDTO> response = couponPolicyBookController.update(1L, couponPolicyBookRequestDTO);
+
+        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
+        assertThat(response.getBody().couponPolicyBookId()).isEqualTo(1L);
+
+        verify(couponPolicyBookService, times(1)).updateCouponPolicyBook(eq(1L), any(CouponPolicyBookRequestDTO.class));
+    }
+
+    @Test
+    public void testDelete() {
+        doNothing().when(couponPolicyBookService).deleteCouponPolicyBook(anyLong());
+
+        ResponseEntity<Void> response = couponPolicyBookController.delete(1L);
+
+        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.NO_CONTENT);
+
+        verify(couponPolicyBookService, times(1)).deleteCouponPolicyBook(1L);
+    }
 }
